Fix display of runs in preview with long names
Consolidate run preview cards and show name, prompt, and model

Fixes #1473
